Fix error deriving changelog when previous_revision is nil
For the first deploy of an application, previous_revision is nil and
thus `fetch(:previous_revision)[/^.*$/]` throws an exception when trying
to run `git log`.

Just omit the changelog on the first deploy of an app since that info
isn't particularly interesting on initial deployment.

For reference, previous_revision is set in
https://github.com/capistrano/capistrano/blob/848301718570b168ef491974e3ba5b5055d63648/lib/capistrano/tasks/deploy.rake#L231-L233